% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsTiMEx.R
\name{analyzePairs}
\alias{analyzePairs}
\title{Finds all mutually exclusive pairs in a dataset}
\usage{
analyzePairs(mat)
}
\arguments{
\item{mat}{binary alteration matrix, with rows representing patients and
columns representing genes}
}
\value{
List consisting of a set of matrices, all of dimension \code{n x n} 
(\code{n} being the number of genes). Element \code{(i,j)} of each matrix 
corresponds to the pairwise test between genes \emph{i} and \emph{j}:
\itemize{
\item{\code{lamiEstNull}} {matrix with rate estimates (lambda) of the 
waiting time of one gene (gene \code{i}) under the Null model.}
\item{\code{lamjEstNull}} {matrix with rate estimates (lambda) of the 
waiting time of the other gene (gene \code{j}) under the Null model.}
\item{\code{lamiEstME}} {matrix with rate estimates (lambda) of the waiting
time of one gene (gene \code{i}) under the Mutual Exclusivity model.}
\item{\code{lamjEstME}} {matrix with rate estimates (lambda) of the 
waiting time of the other gene (gene \code{j}) under the Mutual Exclusivity 
model.}
\item{\code{likeNull}} {matrix with log likelihoods under the Null model.}
\item{\code{likeAlt}} {matrix with log likelihoods under the Mutual
Exclusivity model.}
\item{\code{LRT}} {matrix with log likelihood ratio test (LRT) statistics.}
\item{\code{pvalueLRTCorrectSym}} {list of the following symmetric matrices:
\itemize{
\item{\code{bonferroni}} {bonferroni corrected p-values corresponding to 
the LRT statistic.}
\item{\code{bonferroni}} {fdr corrected p-values corresponding to the 
LRT statistic.}
\item{\code{uncorrected}} {uncorrected  p-values corresponding to the 
LRT statistic.}
}
}
\item{\code{muEstSym}} {symmetric matrix with estimated pairwise mu 
(intensity of mutual exclusivity)}
}
}
\description{
\code{analyzePairs} performs step 1 of the TiMEx procedure:
tests all gene pairs for mutual exclusivity. It returns a complex list, 
including parameter estimates, pairwise p-values and intensities of mutual 
exclusivity, log likelihoods, and others.
}
\details{
In the first step, the TiMEx procedure for identifying mutually
exclusive groups of alterations tests all pairs for mutual exclusivity. The 
data corresponding to each pair of genes is therefore fitted to both the 
Null (Conditional Independence) and the Mutual Exclusivity models. 
Parameters under the two models are estimated, and, since they are nested, 
a likelihood ratio test is performed between the corresponding log 
likelihoods, in order to test whether mu (the intensity of mutual 
exclusivity) is different from 0. For more details on the TiMEx procedure, 
as well as on the underlying mathematical model, see Constantinescu 
\emph{et al.}: \emph{TiMEx: A Waiting Time Model for Mutually Exclusive 
Cancer Alterations}. Bioinformatics (2015).


The output list contains exhaustive information on the pairwise testing of 
genes, including parameter estimates, pairwise p-values and intensities of 
mutual exclusivity, log likelihoods, and others. 

This function displays progress messages indicating the gene that is 
currently being tested against the remaining genes.
}
\examples{
# Test all pairs from the ovarian dataset for mutual exclusivity (takes 
# approximately 5 minutes)
data(ovarian)
\donttest{ovarianPairs<-analyzePairs(ovarian)}

}
\references{
Constantinescu 
\emph{et al.}: \emph{TiMEx: A Waiting Time Model for Mutually Exclusive 
Cancer Alterations}. Bioinformatics (2015).
}
\seealso{
\code{\link{doMaxCliques}} for step 2 of the TiMEx procedure;
\code{\link{findSignifCliques}} for step 3 of the TiMEx procedure;
the wrapper function \code{\link{TiMEx}} for combining these three steps, 
and identifying mutually exclusive groups in a  binary dataset with the 
TiMEx model.
}
\author{
Simona Cristea, \email{scristea@jimmy.harvard.edu}
}
