% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentDatasets.R
\name{breast}
\alias{breast}
\title{Breast cancer dataset}
\format{\code{breast} is a binary matrix with 958 rows and 537 columns.}
\source{
\url{http://www.cbioportal.org/study.do?cancer_study_id=brca_tcga}
}
\description{
Dataset containing a binary alteration pattern for the breast cancer 
dataset downloaded from cBioPortal (TCGA) in July 2014, and preprocessed
as described in Constantinescu \emph{et al.}: \emph{TiMEx: A Waiting Time 
Model for Mutually Exclusive Cancer Alterations}. Bioinformatics (2015). 
Rows represent patients, and columns represent alterations.
}

