% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/documentDatasets.R
\name{breast.rda}
\alias{breast.rda}
\title{Breast cancer dataset}
\format{\code{breast} is a binary matrix with 958 rows and 537 columns.}
\source{
\url{http://www.cbioportal.org/study.do?cancer_study_id=brca_tcga}
}
\description{
Dataset containing a binary alteration pattern for the breast cancer
dataset downloaded from cBioPortal (TCGA) in July 2014, and preprocessed
as explained in "TiMEx: A Waiting Time Model For Mutually
Exclusive Cancer Alterations", by Constantinescu \emph{et al.}
(Bioinformatics, 2015). Rows represent patients, and columns represent
alterations.
}

