% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsTiMEx.R
\name{doMaxCliques}
\alias{doMaxCliques}
\title{Identifies maximal cliques from pairwise testing information}
\usage{
doMaxCliques(pairs, pairMu, pairPvalue)
}
\arguments{
\item{pairs}{list resulting after pairwise testing, as returned by
\code{\link{analyzePairs}}}

\item{pairMu}{pair-level threshold on mu (real number between 0 and 1). 
Default is 0.5.}

\item{pairPvalue}{pair-level threshold on p-value (real number between 0 and
1). Default is 0.01.}
}
\value{
list consisting of:
\itemize{
\item{\code{detectedLengths}} {vector of lengths of the identified maximal 
cliques.}
\item{\code{idxInCliques}} {list with as many elements as lengths of 
the identified maximal cliques. Each element of the list is a matrix, in 
which each row represents the indices of genes in an identified maximal 
clique.}
\item{\code{genesInCliques}} {list with as many elements as lengths of 
the identified maximal cliques. Each element of the list is a matrix, in 
which each row represents the names of genes in an identified maximal 
clique.}
\item{\code{noMaxCliques}} {vector of numbers of identified maximal cliques 
corresponding to each length present in the field \code{detectedLengths}.}
\item{\code{Mus}} {list of two elements: \code{OrderedGenesInCliques} and 
\code{OrderedIdxInCliques}, which have the same structure as the elements
\code{genesInCliques} and \code{idxInCliques}. The only difference is that
the identified maximal cliques are now ordered by their averge pairwise
mu.}
\item \code{pairMu} input pair-level threshold on mu.
\item \code{pairPvalue} input pair-level threshold on p-value.
}
}
\description{
\code{doMaxCliques} performs step 2 of the TiMEx procedure: 
identifies maximal cliques using information from pairwise testing. The 
maximal clique detection routine only uses the connections between gene 
pairs which satisfy the thresholds on mu (\code{pairMu}) and pvalue 
(\code{pairPvalue}).
}
\details{
In the second step, the TiMEx procedure for identifying mutually 
exclusive groups of alterations detects maximal cliques using pairwise 
testing information from step 1. A graph is constructed, in which genes
are vertices, and an edge is drawn between any pair \emph{(i,j)} if both the
estimated intensity of mutual exclusivity and the computed p-value satisfy 
the chosen thresholds \code{pairMu} and \code{pairPvalue}. Maximal
cliques are detected on this graph. 

The two thresholds can be set by the 
user, and are recommended to be chosen based on the sensitivity and 
specificity levels to which they correspond, as assessed in simulated data. 
For details, see 'TiMEx: A Waiting Time Model For Mutually Exclusive Cancer 
Alterations', by Constantinescu \emph{et al.} (2015). The default values are
0.5 for \code{pairMu} and 0.01 for \code{pairPvalue}.

This function needs functions from the packages \emph{RBGL} and 
\emph{igraph} to run.
}
\examples{
# First, test all pairs from the ovarian cancer dataset for mutual 
# exclusivity (take approximately 5 minutes) 
data(ovarian)
\donttest{ovarianPairs<-analyzePairs(ovarian)}

# Then, identify all maximal cliques using the default thresholds
\donttest{ovarianMaxCliques<-doMaxCliques(ovarianPairs)}

}
\author{
Simona Cristea, \email{scristea@jimmy.harvard.edu}
}
\references{
'TiMEx: A Waiting Time Model For Mutually
Exclusive Cancer Alterations', by Constantinescu \emph{et al.} 
(Bioinformatics, 2016).
}
\seealso{
\code{\link{analyzePairs}} for step 1 of the TiMEx procedure;
\code{\link{findSignifCliques}} for step 3 of the TiMEx procedure;
the wrapper function \code{\link{TiMEx}} for combining these three steps, 
and identifying mutually exclusive groups in a  binary dataset with the 
TiMEx model.
}

