% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsTiMEx.R
\name{doMetagene}
\alias{doMetagene}
\title{Creates metagroups of genes}
\usage{
doMetagene(mat)
}
\arguments{
\item{mat}{binary alteration matrix, with rows representing patients and
columns representing genes}
}
\value{
List consisting of:
\itemize{
\item {\code{newMat}} {the collapsed input binary matrix, with metagenes
instead of genes.}
\item {\code{groups}} {list of metagenes, with as many elements as input 
genes which had an identical alteration pattern with at least one other 
input gene.}
}
}
\description{
\code{doMetagene} collapses genes with identical alteration
patterns across patients into metagroups. It returns a new matrix with
the collapsed genes, as well as the members of the metagroups.
}
\details{
It is recommended to run this function on the input binary matrix 
before applying TiMEx, because genes with identical alteration patterns 
across patients will otherwise be indistinguishable.

Note that in the datasets provided in this package, the genes 
have already been collapsed into metagroups.
}
\examples{
# Simulate genes and extract groups
simGenes<-simulateGenes(c(0.5,1,0.3),0.8,4000)
genes<-cbind(simGenes$genes,simGenes$genes)
genesNew<-doMetagene(genes)

# In the datasets provided in this package, the genes have already been 
# collapsed into metagroups, hence the new matrix will be identical to the 
# old one.
data(ovarian)
\donttest{ovarianNew<-doMetagene(ovarian)}

}
\references{
Constantinescu 
\emph{et al.}: \emph{TiMEx: A Waiting Time Model for Mutually Exclusive 
Cancer Alterations}. Bioinformatics (2015).
}
\seealso{
\code{\link{ovarianGroups}}, \code{\link{breastGroups}} 
for examples of metagroups in large cancer datasets.
}
\author{
Simona Cristea, \email{scristea@jimmy.harvard.edu}
}
