% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsTiMEx.R
\name{findSignifCliques}
\alias{findSignifCliques}
\title{Tests all maximal cliques for mutual exclusivity}
\usage{
findSignifCliques(mat, mcStruct, groupPvalue)
}
\arguments{
\item{mat}{binary alteration matrix, with rows representing patients and
columns representing genes}

\item{mcStruct}{list containing maximal cliques, as returned by
\code{\link{doMaxCliques}}}

\item{groupPvalue}{threshold for the corrected p-value of the groups, lower 
than which cliques are significant (real number between 0 and 1). Default 
is 0.1.}
}
\value{
list consisting of:
\itemize{
\item{\code{genesSignif}} {list of significantly mutually exclusive groups, 
as gene names, sorted by corrected p-value. The list contains as many 
elements as identified lengths of groups. For example, 
\code{genesSignif[[2]]}
is a list containing the gene names of the significant groups of size 2. 
Each list of this type further has two elements, \code{fdr} and 
\code{bonf}, corresponding to different multiple testing correction 
methods. Each element is a matrix, in which rows represent gene names of 
significantly mutually exclusive groups.}

\item{\code{idxSignif}} {list of significantly mutually exclusive groups, as
indices in the input matrix, sorted by corrected p-value. The list 
contains as many elements as identified lengths of groups. For example,
\code{idxSignif[[2]]} is a list containing the indices of the  
significant groups of size 2. Each list of this type further has two 
elements, \code{fdr} and \code{bonf}, corresponding to different multiple 
testing correction methods. Each element is a matrix, in which rows 
represent indices of significantly mutually exclusive groups.}

\item{\code{pvals}} {list of corrected significant p-values corresponding to
the tested cliques, ordered ascendingly. The list contains as many elements 
as identified lengths of significant groups. For example, \code{pvals[[2]]}
is a list containing the p-values of the significant maximal cliques of 
size 2. Each list of this type further has two elements, \code{fdr} and 
\code{bonf}, corresponding to different multiple testing correction 
methods. Each element is a vector, of length the number of significant 
maximal cliques of a given size.} 

\item{\code{posSignif}} {list of positions of the significant groups in the 
input list of maximal cliques, ordered ascendingly by corrected p-value. 
The list contains as many elements as identified lengths of significant 
groups. For example, \code{posSignif[[2]]} is a list containing the 
positions of the significant groups of size 2.  Each list of this type 
further has two elements, \code{fdr} and \code{bonf}, corresponding to 
different multiple correction methods.  Each element is a vector, of length 
the number of significant maximal cliques of a given size.}

\item{\code{MusGroup}} {list of inferred mu values corresponding to
the tested cliques, ordered ascendingly by the corresponding corrected 
p-value. The list contains as many elements as identified lengths of 
significant groups. For example, \code{MusGroup[[2]]} is a list containing 
the mu values of the significant maximal cliques of size 2. Each list of 
this type further has two elements, \code{fdr} and \code{bonf}, 
corresponding to different multiple testing correction methods. Each 
element is a vector, of length the number of significant maximal cliques 
of a given size.} 

\item{\code{mcStruct}} {input structure of maximal cliques to be tested 
for mutual exclusivity, as returned by \code{\link{doMaxCliques}}}

\item{\code{matrix}} {input binary alteration matrix}

\item{\code{groupPvalue}} {input threshold for the corrected p-value, lower
than which cliques are significant}
}
}
\description{
\code{findSignifCliques} performs step 3 of the TiMEx 
procedure, namely tests all candidate maximal cliques for mutual 
exclusivity and reports the significant ones.
}
\details{
This function displays progress messages, namely the size of the 
clique currently being tested, and the number of cliques to test. 

Note that sequentially performing steps 1, 2, and 3 of the TiMEx procedure
(functions \code{\link{analyzePairs}}, \code{\link{doMaxCliques}}, and 
\code{\link{findSignifCliques}}) is equivalent to simply running the 
function \code{\link{TiMEx}}.
}
\examples{
# First, test all pairs from the ovarian dataset for mutual exclusivity 
# (takes approximately 5 minutes).
data(ovarian)
\donttest{ovarianPairs<-analyzePairs(ovarian)}

# Second, identify all maximal cliques using the default thresholds
\donttest{ovarianMaxCliques<-doMaxCliques(ovarianPairs)}

# Then, test all maximal cliques for mutual exclusivity and report the 
# significant ones, based on a corrected p-value threshold of 0.1 (default).
\donttest{ovarianMEgroups<-findSignifCliques(ovarian,ovarianMaxCliques)}

}
\author{
Simona Cristea, \email{scristea@jimmy.harvard.edu}
}
\references{
'TiMEx: A Waiting Time Model For Mutually
Exclusive Cancer Alterations', by Constantinescu \emph{et al.} 
(Bioinformatics, 2016).
}
\seealso{
\code{\link{analyzePairs}} for step 1 of the TiMEx procedure;
\code{\link{doMaxCliques}} for step 2 of the TiMEx procedure;
the wrapper function \code{\link{TiMEx}} for combining these three steps, 
and identifying mutually exclusive groups in a  binary dataset with the 
TiMEx model. The data structures \code{\link{ovarianOutput}},
\code{\link{breastOutput}}, \code{\link{gbmDendrixOutput}}, and
\code{\link{gbmMuexOutput}} are examples of structures resulting after 
running TiMEx on large cancer datasets.
}

