% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/documentDatasets.R
\name{gbmDendrix}
\alias{gbmDendrix}
\title{Glioblastoma dataset used by Dendrix}
\format{\code{gbmDendrix} is a binary matrix with 261 rows and 486 columns.}
\source{
\url{http://journals.plos.org/ploscompbiol/article?id=10.1371/
journal.pcbi.1003054}
}
\description{
Dataset containing a binary alteration pattern for the glioblastoma
dataset used by Leiserson \emph{et. al} in "Simultaneous identification of
multiple driver pathways in cancer" (Plos Computational Biology, 2013).
Rows represent patients, and columns represent alterations. In the names
of alterations, \emph{(D)} represents a copy number deletion, and
\emph{(A)} represents a copy number amplification.
}

