% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentDatasets.R
\name{gbmDendrixSubsampling}
\alias{gbmDendrixSubsampling}
\title{Stability of mutually exclusive groups in the glioblastoma datased used 
by Dendrix}
\format{\code{gbmDendrixSubsampling} is a list with as many elements as 
subsampling frequencies provided (3 in this case). Each element is further 
a list with as many elements as number of sizes of the significantly 
mutually exclusive groups identified. Additionally, \code{bonf} and 
\code{fdr} are two lists corresponding to each of these elements, 
representing different multiple correction methods. Finally, each element 
is a vector of relative counts of the significantly mutually exclusive 
groups identified. For example, \code{gbmDendrixSubsampling[[1]][[3]]} 
represents the relative counts of the identified mutually exclusive groups 
of size 3 for a subsampling frequency of 30\%, for both  \code{fdr} and 
\code{bonf} (bonferroni) multiple correction methods.}
\source{
Produced with the function \code{\link{subsampleAnalysis}}, ran with
the inputs 

\code{subsampl<-c(0.3,0.5,0.8)}

\code{noReps<-100}

and the mutually exclusive groups from \code{\link{gbmDendrixOutput}}.
}
\description{
Dataset containing the stability of the mutually exclusive groups identified
by TiMEx in the glioblastoma dataset \code{\link{gbmDendrix}}, used 
in Leiserson \emph{et. al}: \emph{Simultaneous Identification 
of Multiple Driver Pathways in Cancer}. Plos Computational Biology (2013), 
after subsampling the set of patients at frequencies of \code{30\%}, 
\code{50\%}, and \code{80\%}, 100 times.
}
