% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentDatasets.R
\name{gbmMuex}
\alias{gbmMuex}
\title{Glioblastoma dataset used by muex}
\format{\code{gbmMuex} is a binary matrix with 236 rows and 83 columns.}
\source{
\url{http://journals.plos.org/ploscompbiol/article?id=10.1371/
journal.pcbi.1003503}
}
\description{
Dataset containing a binary alteration pattern for the glioblastoma 
dataset used by Szczurek \emph{et. al} in 'Modeling mutual exclusivity of 
cancer mutations' (Research in Computational Molecular Biology, 2014). 
Rows represent patients, and columns represent alterations.
}

