% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/documentDatasets.R
\name{gbmMuexSubsampling.RData}
\alias{gbmMuexSubsampling.RData}
\title{Stability of mutually exclusive groups in the glioblastoma datased used
by muex}
\format{The number of repetitions (\code{noRepsGbmMuex}) and
\code{countsAllGbmMuex}, a list with as many elements as subsampling
frequencies provided. Each element is further a list with as many elemenets
as number of sizes of the significantly mutually exclusive groups
identified. Aditionally, \code{bonf} and \code{fdr} are two lists
corresponding to each of these elements, representing different multiple
correction methods. Finally, each element is a vector of subsampling
frequencies of the significant mutually exclusive groups identified. For
example, \code{countsAllGbmMuex[[1]][[3]]} represents the relative counts
of the identified mutually exclusive groups of size 3 for a subsampling
frequency of 30\%, for both  \code{fdr} and \code{bonf} (bonferroni)
multiple correction methods.}
\source{
Produced with the function \code{\link{subsampleAnalysis}}, ran with
the inputs \code{subsampl<-c(0.3,0.5,0.8)}, \code{noReps<-100}, and the
mutually exclusive groups from \code{\link{gbmMuexOutput.rda}}.
}
\description{
Dataset containing the stability of the mutually exclusive groups identified
by TiMEx in the glioblastoma dataset used by Szczurek \emph{et. al} in
"Modeling mutual exclusivity of cancer mutations" (Research in
Computational Molecular Biology, 2014), after subsampling the set
of patients at frequencies of \code{30\%}, \code{50\%}, and \code{80\%}, for
100 times.
}

