% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentDatasets.R
\name{ovarian}
\alias{ovarian}
\title{Ovarian cancer dataset}
\format{A binary matrix with 316 rows and 312 columns.}
\source{
\url{http://www.cbioportal.org/study.do?cancer_study_id=ov_tcga_pub}
}
\description{
Dataset containing a binary alteration pattern for the ovarian cancer 
dataset downloaded from cBioPortal (TCGA) in July 2014, and preprocessed
as explained in Constantinescu \emph{et al.}: \emph{TiMEx: A Waiting Time 
Model for Mutually Exclusive Cancer Alterations}. Bioinformatics (2015). 
Rows represent patients, and columns represent alterations.
}
