% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentDatasets.R
\name{ovarianGroups}
\alias{ovarianGroups}
\title{Metagroups of genes in ovarian cancer}
\format{\code{ovarianGroups} is a list with 263 elements, where each 
element is a vector of genes with identical alteration patterns as the 
current gene. The numbers indicate the positions of the genes in the input
matrix.}
\source{
Produced with the function \code{\link{doMetagene}}.
}
\description{
Dataset containing the genes with identical alteration patterns in the 
ovarian cancer dataset \code{\link{ovarian}} (before preprocessing). It 
is represented as a list of metagenes, with as many elements as input 
genes which had an identical alteration pattern with at least one other 
input gene.
}

