% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentDatasets.R
\name{ovarianOutput}
\alias{ovarianOutput}
\title{Mutually exclusive groups in ovarian cancer}
\format{\code{ovarianOutput} is a list consisting of:
\itemize{
\item{\code{genesSignif}} {list of significantly mutually exclusive groups, 
as gene names, sorted by corrected p-value. The list contains as many 
elements as identified lengths of groups. For example, 
\code{genesSignif[[2]]} is a list containing the gene names of the 
significant groups of size 2. Each list of this type further has two 
elements, \code{fdr} and \code{bonf}, corresponding to different multiple 
testing correction methods. Each element is a matrix, in which rows 
represent gene names of significantly mutually exclusive groups.}

\item{\code{idxSignif}} {list of significantly mutually exclusive groups, as
indices in the input matrix, sorted by corrected p-value. The list 
contains as many elements as identified lengths of groups. For example,
\code{idxSignif[[2]]} is a list containing the indices of the  
significant groups of size 2. Each list of this type further has two 
elements, \code{fdr} and \code{bonf}, corresponding to different multiple 
testing correction methods. Each element is a matrix, in which rows 
represent indices of significantly mutually exclusive groups.}

\item{\code{pvals}} {list of corrected significant p-values corresponding to
the tested cliques, ordered ascendingly. The list contains as many elements 
as identified lengths of significant groups. For example, \code{pvals[[2]]}
is a list containing the p-values of the significant maximal cliques of 
size 2. Each list of this type further has two elements, \code{fdr} and 
\code{bonf}, corresponding to different multiple testing correction 
methods. Each element is a vector, of length the number of significant 
maximal cliques of a given size.} 

\item{\code{posSignif}} {list of positions of the significant groups in the 
input list of maximal cliques, ordered ascendingly by corrected p-value. 
The list contains as many elements as identified lengths of significant 
groups.For example, \code{posSignif[[2]]} is a list containing the 
positions of the significant groups of size 2.  Each list of this type 
further has two elements, \code{fdr} and \code{bonf}, corresponding to 
different multiple correction methods.  Each element is a vector, of 
length the number of significant maximal cliques of a given size.}

\item{\code{MusGroup}} {list of inferred mu values corresponding to
the tested cliques, ordered ascendingly by the corresponding corrected 
p-value. The list contains as many elements as identified lengths of 
significant groups. For example, \code{MusGroup[[2]]} is a list containing 
the mu values of the significant maximal cliques of size 2. Each list of 
this type further has two elements, \code{fdr} and \code{bonf}, 
corresponding to different multiple testing correction methods. Each 
element is a vector, of length the number of significant maximal cliques 
of a given size.} 

\item{\code{mcStruct}} {input structure of maximal cliques to be tested 
for mutual exclusivity, as returned by \code{\link{doMaxCliques}}.}

\item{\code{matrix}} {input binary alteration matrix.}

\item{\code{groupPvalue}} {input threshold for the corrected p-value, lower
than which cliques are significant.}
}}
\source{
Produced with the function \code{\link{TiMEx}}, on the binary matrix
in the input dataset \code{\link{ovarian}}.
}
\description{
Dataset containing the members of the groups identified as significantly 
mutually exclusive by TiMEx in ovarian cancer, together with their 
intensities of mutual exclusivity, corrected p-values, and other 
information.
}

