% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsTiMEx.R
\name{plotGroupByName}
\alias{plotGroupByName}
\title{Plots a Mutually Exclusive group}
\usage{
plotGroupByName(group, mat)
}
\arguments{
\item{group}{vector of gene names to be plotted}

\item{mat}{binary alteration matrix, with rows representing patients and
columns representing genes}
}
\value{
None
}
\description{
\code{plotGroupByName} plots a mutually exclusive group 
(including the frequencies of the genes), given the names of the genes in 
the group.
}
\details{
The plotting is done based on the function 
\code{\link[graphics]{image}}.
}
\examples{
# Plot the group consisting of the copy number aberrations of MIEN1 and 
# CDKN1B, and the point mutations of CDH1, GATA3, and MAP3K1, in breast 
# cancer.
data(breast)
group<-c('MIEN1-CNA','CDH1-Mut','GATA3-Mut','MAP3K1-Mut','CDKN1B-CNA')
plotGroupByName(group,breast)

}
\author{
Simona Cristea, \email{scristea@jimmy.harvard.edu}
}
\references{
'TiMEx: A Waiting Time Model For Mutually
Exclusive Cancer Alterations', by Constantinescu \emph{et al.} 
(Bioinformatics, 2016).
}
\seealso{
the wrapper function \code{\link{TiMEx}} for identifying
mutually exclusive groups in a  binary dataset with the TiMEx model.
}

