% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internSimulations.R
\name{produceGenesGroup}
\alias{produceGenesGroup}
\title{(internal) main function to simulate genes corresponding to the 
probability distribution in given as input genoMat (n=number of genes; 
NCur=sample size)}
\usage{
produceGenesGroup(NCur, genoMat, n)
}
\description{
(internal) main function to simulate genes corresponding to the 
probability distribution in given as input genoMat (n=number of genes; 
NCur=sample size)
}
