% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functionsTiMEx.R
\name{produceTablesSignifGroups}
\alias{produceTablesSignifGroups}
\title{Produces tables with groups}
\usage{
produceTablesSignifGroups(signifGroups, mat, noToShow)
}
\arguments{
\item{signifGroups}{result structure with the significant groups, as
returned by either \code{\link{TiMEx}} or \code{\link{findSignifCliques}}}

\item{mat}{binary alteration matrix, with rows representing patients and
columns representing genes}

\item{noToShow}{maximum number of groups to include in the table. Default
is 30.}
}
\value{
list with as many elements as lengths of the identified mutually
exclusive groups, containing tables with the significant groups for each
size.
Each list of this type further has two elements, \code{fdr} and
\code{bonf}, corresponding to different multiple testing correction
methods. Each element is a matrix, in which rows represent significantly
mutually exclusive groups.
}
\description{
\code{produceTablesSignifGroups} produces tables with the
significant groups. These tables include the names of the genes part of the
groups, their respective frequency in the dataset, and the mu and corrected
pvalue corresponding to each group.
}
\details{
This function summarizes information on the mutually exclusive
groups identified by TiMEx in a dataset, as tables.
}
\examples{
# Produce tables on the output of TiMEx on the ovarian cancer dataset
data(ovarian)
data(ovarianOutput)
ovarianTables<-produceTablesSignifGroups(ovarianOutput,ovarian)
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}
\references{
"TiMEx: A Waiting Time Model For Mutually
Exclusive Cancer Alterations", by Constantinescu \emph{et al.}
(Bioinformatics, 2015).
}
\seealso{
the wrapper function \code{\link{TiMEx}} for identifying
mutually exclusive groups in a  binary dataset with the TiMEx model.
}

