% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsTiMEx.R
\name{recoverAllNamesGroups}
\alias{recoverAllNamesGroups}
\title{Recovers members of the metagroups}
\usage{
recoverAllNamesGroups(groupsMeta, clGenes)
}
\arguments{
\item{groupsMeta}{list containing groups of equivalent genes, as returned 
by the field \code{groups} of \code{\link{doMetagene}}}

\item{clGenes}{matrix of mutually exclusive groups of same size, as gene 
names. This type of matrix is returned by either \code{\link{TiMEx}} or
\code{\link{findSignifCliques}}, as one of the 
matrix elements of the \code{genesSignif} field.}
}
\value{
list with as many elements as number of identified mutually 
exclusive groups, \emph{i.e.} number of rows in the input matrix. Each of 
its elements is further a list, containing, at each position, the metagroup 
of the genes in the initial group at that resepective position. For an 
example, see \emph{Details} above.
}
\description{
\code{recoverAllNameGroups}  recovers, from a metagroup, the 
names of the genes part of an identified mutually exclusive group.
}
\details{
This function can be used if the input binary matrix contains 
identical events that need to be merged into metagenes using 
\code{\link{doMetagene}}. Running \code{recoverAllNamesGroups} provides the
set of identical alterations which are part of the identified mutually 
exclusive groups. 

In order to run this function on all the identified mutually exclusive 
groups as returned by \code{\link{TiMEx}} or 
\code{\link{findSignifCliques}}, it is necessary to run it separately on 
each matrix element (corresponding to different group sizes and different 
correction methods) of the \code{genesSignif} field in the structure 
returned by either \code{\link{TiMEx}} or \code{\link{findSignifCliques}}.

For example, after loading \code{data(ovarianGroups)} and 
\code{data(ovarianOutput)}, and running 

\code{rGroups<-recoverAllNamesGroups(ovarianGroups,
signifGroups$genesSignif[[3]]$bonf)}

\code{rGroups[[14]]} has 3 elements (as many as genes part
of the identified mutually exclusive groups). Each element is the metagroup 
of each of the genes part of the 14th mutually exclusive group in the input
matrix 
\code{signifGroups$genesSignif[[3]]$bonf}. Namely, \emph{BRD4-CNA} and 
\emph{MYC-CNA} have unique alteration patterns among samples, and are alone 
in their metagroup, while \emph{CASC1-CNA} has an identical alteration
pattern with \emph{KRAS-CNA} and \emph{LYRM5-CNA}. The numbers
below the gene names are the indices of the genes in the initial input 
binary matrix of patients.
}
\examples{
data(ovarianGroups)
data(ovarianOutput)
r<-recoverAllNamesGroups(ovarianGroups,ovarianOutput$genesSignif[[3]]$bonf)

}
\author{
Simona Cristea, \email{scristea@jimmy.harvard.edu}
}
\references{
Constantinescu 
\emph{et al.}: \emph{TiMEx: A Waiting Time Model for Mutually Exclusive 
Cancer Alterations}. Bioinformatics (2015).
}
\seealso{
\code{\link{doMetagene}} for collapsing the genes of an input 
matrix with identical alteration patterns into metagroups.
}

