% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsTiMEx.R
\name{removeLowFreqs}
\alias{removeLowFreqs}
\title{Removes alterations based on frequency}
\usage{
removeLowFreqs(mat, level)
}
\arguments{
\item{mat}{binary alteration matrix, with rows representing patients and
columns representing genes}

\item{level}{frequency level under which the genes are to be removed 
(real number between 0 and 1). Default is 0.03.}
}
\value{
the binary input matrix without the low-frequently altered genes.
}
\description{
\code{removeLowFreqs} returns a binary matrix from which genes 
altered with lower frequency than the input level are removed.
}
\details{
It is only recommended to run this function if the user is not at 
all interested in the role of low-frequently altered genes.
}
\examples{
# Remove genes altered in less than 3\% (the default level) of the samples  
# in the ovarian cancer dataset
data(ovarian)
ovarianNew<-removeLowFreqs(ovarian)

}
\author{
Simona Cristea, \email{scristea@jimmy.harvard.edu}
}
\references{
Constantinescu 
\emph{et al.}: \emph{TiMEx: A Waiting Time Model for Mutually Exclusive 
Cancer Alterations}. Bioinformatics (2015).
}
\seealso{
\code{\link{ovarian}}, \code{\link{breast}}, 
and \code{\link{gbmDendrix}}  for examples of biological large cancer
datasets.
}

