% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsTiMEx.R
\name{simulateGenes}
\alias{simulateGenes}
\title{Generates data from the TiMEx model}
\usage{
simulateGenes(lambdas, mu, N)
}
\arguments{
\item{lambdas}{vector of exponential rates (positive real numbers). 
The length of the vector equals the number of simulated genes.}

\item{mu}{intensity of mutual exclusivity (real number between 
0 and 1). Default is 1 (perfect mutual exclusivity).}

\item{N}{sample size (positive integer).}
}
\value{
list consisting of:
\itemize{
\item{\code{genes}} {the simulated dataset, as a binary matrix.}
\item{\code{genoMat}} {the matrix with genotype probabilities, from which 
the dataset was simulated. This matrix has as many dimensions as number of 
genes, i.e. 2x2x...x2. For each dimension, the first position corresponds to
the probability of observing a 0 for that gene, and the second position 
corresponds to the probability of observing an 1. For example, in the case 
of 4 genes, the probability of observing the null genotype \code{(0000)} is 
given by \code{genoMat[1,1,1,1]}; the probability of observing the genotype 
\code{(1011)} is given by \code{genoMat[2,1,2,2]}; the probability of 
observing the genotype \code{(1111)} is given by \code{genoMat[2,2,2,2]}. 
The entries of this matrix are nonnegative and sum up to 1.}
}
}
\description{
\code{simulateGenes} returns a list containing a binary matrix 
simulated  from the TiMEx model, for given lambdas (exponential rates), 
mu (intensity of mutual exclusivity), and N (sample size).
}
\details{
This function needs \code{\link[gtools]{permutations}} in order to 
run.

For details on how the values of the exponential rates correspond to 
frequencies, see \emph{References} below.
}
\examples{
simGenes<-simulateGenes(c(0.5,1,0.3),0.8,4000)

}
\author{
Simona Cristea, \email{scristea@jimmy.harvard.edu}
}
\references{
Constantinescu 
\emph{et al.}: \emph{TiMEx: A Waiting Time Model for Mutually Exclusive 
Cancer Alterations}. Bioinformatics (2015).
}
\seealso{
the wrapper function \code{\link{TiMEx}} for identifying
mutually exclusive groups in a  binary dataset with the TiMEx model,
\code{\link{ovarian}}, \code{\link{breast}}, and
\code{\link{gbmDendrix}}  for examples of biological large cancer 
datasets.
}

