% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsTiMEx.R
\name{subsampleAnalysis}
\alias{subsampleAnalysis}
\title{Assesses the stability of groups by subsampling}
\usage{
subsampleAnalysis(subsampl, noReps, signifGroups)
}
\arguments{
\item{subsampl}{a vector with subsampling frequencies}

\item{noReps}{number of repetitions of subsampling. Default is 100.}

\item{signifGroups}{result structure with the significant groups, as 
returned by either \code{\link{TiMEx}} or \code{\link{findSignifCliques}}}
}
\value{
list with as many elements as subsampling frequencies provided. Each
element is further a list with as many elemenets as number of sizes of the
significantly mutually exclusive groups identified. Aditionally, \code{bonf}
and \code{fdr} are two lists corresponding to each of these elements, 
representing different multiple correction methods. Finally, each element 
is a vector of subsampling frequencies of the significantly mutually 
exclusive groups identified. For an example, see \emph{Details} above.
}
\description{
\code{subsampleAnalysis} subsamples the set of patients and 
assess the stability of the identified mutually exclusive groups.
}
\details{
As this function runs TiMEx many times sequentially, it is 
computationally very intensive. For a version of this function which can be
directly ran on a cluster, please contact me (see e-mail below).

For example, after loading \code{data(ovarianOutput)} and running 

\code{counts<-subsampleAnalysis(subsampl=c(0.3,0.5,0.8),
signifGroups=signifGroups)}

\code{counts[[1]][[3]]} will represent
the relative counts of the identified mutually exclusive groups of size 3 
for a subsampling frequency of 30\%, for both  \code{fdr} and \code{bonf} 
(bonferroni) multiple correction methods.
}
\examples{
# running this function is time-intensive
data(ovarianOutput)
\dontrun{subsampleOvarian<-subsampleAnalysis(c(0.3,0.5,0.8),ovarianOutput)}

}
\author{
Simona Cristea, \email{scristea@jimmy.harvard.edu}
}
\references{
'TiMEx: A Waiting Time Model For Mutually
Exclusive Cancer Alterations', by Constantinescu \emph{et al.} 
(Bioinformatics, 2016).
}
\seealso{
the wrapper function \code{\link{TiMEx}} for identifying mutually 
exclusive groups in a  binary dataset with the TiMEx model,
\code{\link{ovarianSubsampling}}, 
\code{\link{breastSubsampling}}, 
\code{\link{gbmDendrixSubsampling}}, and 
\code{\link{gbmMuexSubsampling}} for examples of outputs after
performing the subsampling analysis.
}

