% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsTiMEx.R
\name{testCliqueAsGroup}
\alias{testCliqueAsGroup}
\alias{testCliquesAsGroup}
\title{Tests whether a given group is mutually exclusive}
\usage{
testCliqueAsGroup(geneIdx, mat, lo)
}
\arguments{
\item{geneIdx}{vector of indices in the input matrix of the genes to be 
tested}

\item{mat}{binary alteration matrix, with rows representing patients and
columns representing genes}

\item{lo}{rate of observation time. Default is 1.}
}
\value{
List consisting of:
\itemize{
\item{\code{opMu}} {list as returned by \code{\link[stats]{optim}}. The 
field \code{par} is a vector of \code{n+1} positions (\code{n} being the 
number of genes) containing the estimates of the waiting time rates 
(lambda) for the \code{n} genes under the mutual exclusivity model, 
followed by the estimate for mu.}
\item{\code{opNull}} {list as returned by \code{\link[stats]{optim}}. The 
field \code{par} is a vector of \code{n} positions (\code{n} being the 
number of genes) containing the estimates of the waiting time rates 
(lambda) for the \code{n} genes under the null model.}
\item{\code{countsVec}} {the contingency table of the input genes, as a 
vector. The first element is the count of the samples where no gene was 
altered, the next \code{n} elements are the counts of the samples where 
exactly one gene was altered (starting with gene 1 and ending with gene 
\code{n}), the next \code{n} elements are the counts of the samples where
exactly two genes were altered (starting with genes 1 and 2, continuing with
genes 1 and 3, and ending with genes \code{n-1} and \code{n}), and so on.}
\item{\code{genes}} {subset of the input binary matrix, corresponding to the
genes to be tested}
\item{\code{LRT}} {log likelihood ratio (LRT)}
\item{\code{pvalueLRT}} {the p-value corresponding to the LRT}
}
}
\description{
\code{testCliqueAsGroup} tests whether a group, given as gene
indices, is mutually exclusive.
}
\details{
For deciding whether a group is mutually exclusive, the group of  
genes is fitted to both the Null (Conditional Independence) and the Mutual 
Exclusivity models. Parameters under the two models are estimated, and, 
since they are nested, a likelihood ratio test is performed between the 
corresponding log likelihoods, in order to test whether mu (the intensity of
mutual exclusivity) is different from 0. For computing the likelihood of 
the data under both models, an exhaustive enumeration of all possible 
orders of the input alterations needs to be performed. Therefore, the 
complexity of the test is exponential in the number of genes to be tested, 
which makes it unfeasible for large number of genes (usually more than 10).

\code{lo} (the rate of observation time) is by default set to 1, as both 
models are otherwise unindentifiable. We recommend leaving the value of this
parameter unchanged, as otherwise the estimated waiting time rates of the
genes require additional interpretation.

For more details on the TiMEx procedure, as well as on the underlying 
mathematical model, see Constantinescu 
\emph{et al.}: \emph{TiMEx: A Waiting Time Model for Mutually Exclusive 
Cancer Alterations}. Bioinformatics (2015).
}
\examples{
# Tests for mutual exclusivity the group of genes with indices 13, 204, and 
# 310 in the ovarian cancer dataset
data(ovarian)
testGroup<-testCliqueAsGroup(c(13,204,310),ovarian)

}
\references{
Constantinescu 
\emph{et al.}: \emph{TiMEx: A Waiting Time Model for Mutually Exclusive 
Cancer Alterations}. Bioinformatics (2015).
}
\seealso{
the wrapper function \code{\link{TiMEx}} for identifying
mutually exclusive groups in a  binary dataset with the TiMEx model.
}
\author{
Simona Cristea, \email{scristea@jimmy.harvard.edu}
}
