% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsTiMEx.R
\name{TiMEx}
\alias{TiMEx}
\title{Finds mutually exclusive groups}
\usage{
TiMEx(mat, pairMu, pairPvalue, groupPvalue)
}
\arguments{
\item{mat}{binary alteration matrix, with rows representing patients and
columns representing genes}

\item{pairMu}{pair-level threshold on mu (real number between 0 and 1). 
Default is 0.5.}

\item{pairPvalue}{pair-level threshold on p-value (real number between 0 
and 1). Default is 0.01.}

\item{groupPvalue}{threshold for the corrected p-value, lower than which 
cliques are significant. Default to 0.1}
}
\value{
list consisting of:
\itemize{
\item{\code{genesSignif}} {list of significantly mutually exclusive groups, 
as gene names, sorted by corrected p-value. The list contains as many 
elements as identified lengths of groups. For example, 
\code{genesSignif[[2]]} is a list containing the gene names of the 
significant groups of size 2. Each list of this type further has two 
elements, \code{fdr} and \code{bonf}, corresponding to different multiple 
testing correction methods. Each element is a matrix, in which rows 
represent gene names of significantly mutually exclusive groups.}

\item{\code{idxSignif}} {list of significantly mutually exclusive groups, 
as indices in the input matrix, sorted by corrected p-value. The list 
contains  as many elements as identified lengths of groups. For example,
\code{idxSignif[[2]]} is a list containing the indices of the  
significant groups of size 2. Each list of this type further has two 
elements, \code{fdr} and \code{bonf}, corresponding to different multiple 
testing correction methods. Each element is a matrix, in which rows 
represent indices of significantly mutually exclusive groups.}

\item{\code{pvals}} {list of corrected significant p-values corresponding to
the tested cliques, ordered ascendingly. The list contains as many elements 
as identified lengths of significant groups. For example, \code{pvals[[2]]}
is a list containing the p-values of the significant maximal cliques of 
size 2. Each list of this type further has two elements, \code{fdr} and 
\code{bonf}, corresponding to different multiple testing correction 
methods. Each element is a vector, of length the number of significant 
maximal cliques of a given size.} 

\item{\code{posSignif}} {list of positions of the significant groups in the 
input list of maximal cliques, ordered ascendingly by corrected p-value. 
The list contains as many elements as identified lengths of significant 
groups. For example, \code{posSignif[[2]]} is a list containing the 
positions of the significant groups of size 2.  Each list of this type 
further has two elements, \code{fdr} and \code{bonf}, corresponding to 
different multiple correction methods.  Each element is a vector, of 
length the number of significant maximal cliques of a given size.}

\item{\code{MusGroup}} {list of inferred mu values corresponding to
the tested cliques, ordered ascendingly by the corresponding corrected 
p-value. The list contains as many elements as identified lengths of 
significant groups. For example, \code{MusGroup[[2]]} is a list containing 
the mu values of the significant maximal cliques of size 2. Each list of 
this type further has two elements, \code{fdr} and \code{bonf}, 
corresponding to different multiple testing correction methods. Each 
element is a vector, of length the number of significant maximal cliques of 
a given size.} 

\item{\code{mcStruct}} {input structure of maximal cliques to be tested 
for mutual exclusivity, as returned by \code{\link{doMaxCliques}}.}

\item{\code{matrix}} {input binary alteration matrix.}

\item{\code{groupPvalue}} {input threshold for the corrected p-value, lower
than which cliques are significant.}
}
}
\description{
\code{TiMEx} is the main function of this package. It 
identifies all groups of mutually exclusive cancer alterations in a large 
binary input dataset.
}
\details{
Dependning on the size of the dataset (both in terms of samples 
and alterations), TiMEx can require a reasonable time to run. For example, 
the approximate running time is 10 minutes for the 
\code{\link{ovarian}} cancer dataset, and 45 minutes for the 
\code{\link{breast}} cancer  dataset included in this package, on a 
personal computer.

\code{TiMEx} displays progress messages. In a fist step, it indicates
the gene which is currently being tested against the remaining genes. In a 
later step, it indicates the size of the clique currently being tested, 
and the number of cliques to test.
}
\examples{
# Run TiMEx on the ovarian cancer dataset with default parameters 
# (takes approximately 10 minutes)
data(ovarian)
\donttest{ovarianMEGroups<-TiMEx(ovarian)}

}
\author{
Simona Cristea, \email{scristea@jimmy.harvard.edu}
}
\references{
'TiMEx: A Waiting Time Model For Mutually
Exclusive Cancer Alterations', by Constantinescu \emph{et al.} 
(Bioinformatics, 2016).
}
\seealso{
\code{\link{analyzePairs}} for step 1 of the TiMEx procedure;
\code{\link{doMaxCliques}} for step 2 of the TiMEx procedure, and 
\code{\link{findSignifCliques}} for step 3 of the TiMEx procedure. The data
structures \code{\link{ovarianOutput}}, \code{\link{breastOutput}},
\code{\link{gbmDendrixOutput}}, and \code{\link{gbmMuexOutput}} 
are examples of structures resulting after running TiMEx on large cancer 
datasets.
}

