% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentDatasets.R
\name{breastSubtypes}
\alias{breastSubtypes}
\title{Breast cancer subtypes}
\format{\code{breastSubtypes} is a list with 4 elements, corresponding to 
the 4 breast subtypes. Each element is binary matrix with 537 columns, as 
follows: 
\code{breastSubtypes$luminalA} consists of 222 rows,
\code{breastSubtypes$luminalB} consists of 125 rows,
\code{Her2} consists of 55 rows, and \code{Basal} consists of 76 rows.}
\source{
\url{http://www.cbioportal.org/study.do?cancer_study_id=brca_tcga}
}
\description{
Dataset containing binary alteration patterns for the breast cancer subtypes
'luminalA', 'luminalB', 'Her2', and 'Basal2', downloaded from cBioPortal 
(TCGA) in July 2014, and preprocessed as explained in 'TiMEx: A Waiting 
Time Model For Mutually Exclusive Cancer Alterations', by Constantinescu 
\emph{et al.}  (Bioinformatics, 2016). Rows represent patients, and columns 
represent alterations.
}

